feature: usePlayText 훅 분리 & 기타 리팩터링
